pyhton and flask grammars and rules
i added the python and flask's grammars and rules by creating a lexer and a parser for the two.
also i added some files to build the project's structure like the AST folder which will have the AST nodes classes soon.
the whole work until now is inside the CompilerPractice folder.