Added AST Nodes!
using the structure i prepared earlier for the ast nodes, i created the ast core nodes that don't need to be duplicated in order to be used for both (html , css and jinja2) or (python and flask) just like the binary and unary operations that are the same for all those languages and frameworks, and also the ast python nodes that will only be used for the operations that are different between the 2 grammars like the for loops and the if/else conditions.
that's all for now.