Major edits
there were some errors when trying to run the grammars apparently because of the wrong generation way i have used before.
made a test file to try everything out and finally they're all working!
also added while loop and try block just to extend my grammars a little bit.
i have made a label table cause i read that it might be useful in the visitor and i'll see haw can i benefit from it for the project.
also the old symbol table was kinda small so i wanted to make it better adding a new class to be the symbol row and the table for having scopes entering and exiting functions.
next step will be ensuring the ast nodes i created were enough and start building the visitor.