AST nodes changes
i edited the ast nodes to fulfill our grammars needs in order to build a proper visitor. ast visitor is just an interface to include a visiting method for each needed node in this language.
nodes are being separated and sorted for core nodes that might be used for both the python_flask grammars side and the html_css_jinja2 side, the others are for each language itself only. (flask is inside python's folder.