AST nodes edits and Symbol Table visitor Building.
built the symbol table visitor, just a simple class.
added the ast visitor interface for the html, css and jinja nodes and implemented the accept method for each node in the project with a null in the return for the html accept visitors in the python nodes, and a visit to the html node if it is a core node.
the python builder visitor is nearly done but there are some indentation errors are occurring after the full implementation of all the nodes for the rules created until this moment. the problem is still not fixed so i didn't want to publish the visitor until we make sure the errors are not inside it.
if it took a while we'll publish it and continue working on the errors after that but we'll give it a day or 2 of our time first.
Note : some nodes' names has been changed and others are reshaped completely, while other needed nodes were missing so they're created in this push. ASTVisitor and HtmlWithCssVisitor are the names of the interfaces.