Added ast nodes from the python branch and fixed the html nodes
Got every AST node from the other branch, including the abstract AST node, so i had to change the constructor of every other node that is html or css or jinja.
some jinja was not being parsed right, so i fixed it in the g4 files. i started working on the visitor but  it was not suitable so i did not upload it