Huge changes!
we extended the grammars g4 files we have, new ways of indentation handling, using channel(hidden) and some other functions, and some order changes in the parser in addition to a better handling for the lists, decorators, try blocks and logical expressions in general, with the generated files being pushed with the code.
some ast changes made by the html , css and jinja2 side of the project but needed here cause it is intersected with this grammars, like the ast visitor, symbol and label tables, and their visitor, and even the ast base visitor class.
the visitor builder is being checked before pushing it by our teammate who will push it this day soon with the test files.